requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage==3.6.0
Flask==3.1.2
google-cloud-bigquery==3.38.0
gunicorn==23.0.0
google-auth==2.43.0
Werkzeug==3.1.3
python-dotenv==1.2.1
beautifulsoup4
requests
pandas-gbq